+++ src/js/utils.js
@@ -47,6 +47,16 @@ function ShowTag(tag) {
         id.className="show";
     }
 }
+function FlipTag(tag) {
+    var id=document.getElementById(tag);
+    if (id) {
+        if  (id.className == "show") {
+            id.className="hide"; 
+        } else {
+            id.className="show"; 
+        }
+    }
+}
 function load(url) {
     window.location.href=url;
 }