+++ src/python/DAS/web/help_cards.py
@@ -15,8 +15,8 @@ def help_cards(base):
     Return list of help cards used on web UI
     """
     first = """
-<h3 class="big">DAS queries</h3>
-If you new to DAS try to search desired data by using 
+<h3 class="big">Help: DAS queries</h3>
+DAS queries are formed by
 <b>key=value</b> pairs, for example
 <ul>
 <li>dataset=*RelVal*</li>
@@ -34,7 +34,7 @@ For more details please read DAS
     card_list = []
 
     card = """
-<h3 class="big">DAS filters</h3>
+<h3 class="big">Help: DAS filters</h3>
 Do you know that DAS supports filters? Similar to UNIX you can apply
 <b>grep</b> filter to select certain data. For example,
 if you look for a given dataset=*RelVal*, you can select its number of events
@@ -50,7 +50,7 @@ dataset=*RelVal* | grep dataset.name, dataset.nevents, dataset.nfiles
     card_list.append(card)
 
     card = """
-<h3 class="big">DAS filters and conditions</h3>
+<h3 class="big">Help: DAS filters and conditions</h3>
 Do you know that DAS filters supports conditions. Their usage is trivial:
 <pre>
 file dataset=/a/b/c | grep file.name, file.size>3000000, file.size<6000000
@@ -65,7 +65,7 @@ dataset=*RelVal* | grep dataset.name, dataset.name=*RECO
     card_list.append(card)
 
     card = """
-<h3 class="big">DAS aggregators</h3>
+<h3 class="big">Help: DAS aggregators</h3>
 DAS supports variety of aggregator functions, such as:
 <b>min, max, sum, count, avg, median</b>. They can be applied in any
 order to any DAS record attribute. For example:
@@ -78,7 +78,7 @@ Custom map-reduce function are also supported. Please contact DAS
     card_list.append(card)
 
     card = """
-<h3 class="big">DAS date usage</h3>
+<h3 class="big">Help: DAS date usage</h3>
 DAS has a special <b>date</b> keyword. It can accepts values either in
 YYYYMMDD format or via <b>last</b> operator. For example:
 <pre>
@@ -92,7 +92,7 @@ Supported units for <b>last</b> operator are <b>d</b> (days),
     card_list.append(card)
 
     card = """
-<h3 class="big">DAS special keywords</h3>
+<h3 class="big">Help: DAS special keywords</h3>
 DAS has special <b>records</b> and <b>queries</b> keywords.
 The former can be used to look-up all records in DAS cache, while
 later shows most recent queries placed into DAS. The DAS filters
@@ -105,7 +105,7 @@ records | grep file.name
     card_list.append(card)
 
     card = """
-<h3 class="big">DAS command line interface</h3>
+<h3 class="big">Help: DAS command line interface</h3>
 Do you know that you can use DAS from your terminal. Go to
 <b><a href="%s/cli">CLI</a></b> link and save it on your disk, e.g. as <em>das_cli</em>.
 Then you can use it as simple as
@@ -124,7 +124,7 @@ python das_cli --query="dataset=/ExpressPhysics* | grep dataset.name, dataset.ne
     card_list.append(card)
 
     card = """
-<h3 class="big">DAS colors</h3>
+<h3 class="big">Help: DAS colors</h3>
 Do you know that colored little box, e.g.
 <cite style="background-color:%s;padding-left:7px; padding-right:7px">&nbsp;</cite>
 next to the Record output